refactor: Change the path of 'SPH' related files to 'Particle'
