+++ Includes/Core/Particle/SPH/SPHStdKernel3.hpp
