+++ Sources/API/Python/Particle/SPH/SPHSystemData.cpp
@@ -8,9 +8,9 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <API/Python/SPH/SPHSystemData.hpp>
-#include <Core/SPH/SPHSystemData2.hpp>
-#include <Core/SPH/SPHSystemData3.hpp>
+#include <API/Python/Particle/SPH/SPHSystemData.hpp>
+#include <Core/Particle/SPH/SPHSystemData2.hpp>
+#include <Core/Particle/SPH/SPHSystemData3.hpp>
 
 #include <pybind11/pybind11.h>
