+++ Sources/Core/Particle/SPH/SPHStdKernel3.cpp
@@ -13,7 +13,7 @@
 // "Meshless Approximation Methods and Applications in Physics Based Modeling
 // and Animation", Eurographics 2009 Tutorial
 
-#include <Core/SPH/SPHStdKernel3.hpp>
+#include <Core/Particle/SPH/SPHStdKernel3.hpp>
 #include <Core/Utils/Constants.hpp>
 
 namespace CubbyFlow
