+++ Sources/Core/PointsToImplicit/AnisotropicPointsToImplicit2.cpp
@@ -10,8 +10,8 @@
 
 #include <Core/Math/SVD.hpp>
 #include <Core/Matrix/Matrix2x2.hpp>
+#include <Core/Particle/SPH/SPHSystemData2.hpp>
 #include <Core/PointsToImplicit/AnisotropicPointsToImplicit2.hpp>
-#include <Core/SPH/SPHSystemData2.hpp>
 #include <Core/Searcher/PointKdTreeSearcher2.hpp>
 #include <Core/Solver/LevelSet/FMMLevelSetSolver2.hpp>
 #include <Core/Utils/Logging.hpp>
