+++ Sources/Core/PointsToImplicit/SPHPointsToImplicit2.cpp
@@ -8,8 +8,8 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
+#include <Core/Particle/SPH/SPHSystemData2.hpp>
 #include <Core/PointsToImplicit/SPHPointsToImplicit2.hpp>
-#include <Core/SPH/SPHSystemData2.hpp>
 #include <Core/Solver/LevelSet/FMMLevelSetSolver2.hpp>
 #include <Core/Utils/Logging.hpp>
