+++ Sources/Core/PointsToImplicit/SPHPointsToImplicit3.cpp
@@ -8,8 +8,8 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
+#include <Core/Particle/SPH/SPHSystemData3.hpp>
 #include <Core/PointsToImplicit/SPHPointsToImplicit3.hpp>
-#include <Core/SPH/SPHSystemData3.hpp>
 #include <Core/Solver/LevelSet/FMMLevelSetSolver3.hpp>
 #include <Core/Utils/Logging.hpp>
