+++ Sources/Core/Solver/Particle/PCISPH/PCISPHSolver2.cpp
@@ -8,8 +8,8 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
+#include <Core/Particle/SPH/SPHStdKernel2.hpp>
 #include <Core/PointGenerator/TrianglePointGenerator.hpp>
-#include <Core/SPH/SPHStdKernel2.hpp>
 #include <Core/Solver/Particle/PCISPH/PCISPHSolver2.hpp>
 #include <Core/Utils/Logging.hpp>
